# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Teleasistencia Documentation",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/avisos": {
            "get": {
                "tags": [
                    "Avisos"
                ],
                "summary": "Listar todos los avisos y alarmas con filtros",
                "operationId": "0079ac224443891e2d55aa3937248b1c",
                "parameters": [
                    {
                        "name": "type",
                        "in": "query",
                        "description": "Filtrar por tipo (alert/alarm)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "alert",
                                "alarm"
                            ]
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filtrar por estado",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "pending",
                                "in_progress",
                                "resolved"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de avisos y alarmas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/AvisoResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Avisos"
                ],
                "summary": "Crear un nuevo aviso o alarma",
                "operationId": "4287ab19168ddfc8527f258f587bab3b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreAlertRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Aviso o alarma creada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/AvisoResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/avisos/{id}": {
            "get": {
                "tags": [
                    "Avisos"
                ],
                "summary": "Obtener detalles de un aviso o alarma",
                "operationId": "90a3cd0d853c5bdc4a033206ae38754d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del aviso o alarma",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalles del aviso o alarma recuperados exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AvisoResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Aviso o alarma no encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Avisos"
                ],
                "summary": "Actualizar un aviso o alarma existente",
                "operationId": "09e10f1188bfc72bb17d9d5463df3efe",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del aviso o alarma",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateAlertRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Aviso o alarma actualizado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AvisoResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Aviso o alarma no encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Avisos"
                ],
                "summary": "Eliminar un aviso o alarma",
                "operationId": "1487e3611c34606e2ddbdbf0c19ee19d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del aviso o alarma",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Aviso o alarma eliminado exitosamente"
                    },
                    "404": {
                        "description": "Aviso o alarma no encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "StoreAlertRequest": {
                "description": "Validación para la creación y actualización de alertas",
                "required": [
                    "tipo",
                    "fecha_hora",
                    "descripcion",
                    "completado",
                    "categoria_id",
                    "paciente_id",
                    "operador_id"
                ],
                "properties": {
                    "id": {
                        "description": "ID de la alerta existente (opcional)",
                        "type": "integer",
                        "example": 1,
                        "nullable": true
                    },
                    "tipo": {
                        "description": "Tipo de alerta",
                        "type": "string",
                        "enum": [
                            "puntual",
                            "periodico"
                        ],
                        "example": "puntual"
                    },
                    "fecha_hora": {
                        "description": "Fecha y hora de la alerta",
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-02-24T01:45:22Z"
                    },
                    "descripcion": {
                        "description": "Descripción detallada de la alerta",
                        "type": "string",
                        "example": "Paciente necesita atención inmediata"
                    },
                    "completado": {
                        "description": "Estado de la alerta",
                        "type": "boolean",
                        "example": false
                    },
                    "fecha_completado": {
                        "description": "Fecha y hora en que la alerta fue completada",
                        "type": "string",
                        "format": "date-time",
                        "example": null,
                        "nullable": true
                    },
                    "categoria_id": {
                        "description": "ID de la categoría de la alerta",
                        "type": "integer",
                        "example": 2
                    },
                    "paciente_id": {
                        "description": "ID del paciente asociado a la alerta",
                        "type": "integer",
                        "example": 5
                    },
                    "operador_id": {
                        "description": "ID del operador responsable de la alerta",
                        "type": "integer",
                        "example": 3
                    }
                },
                "type": "object"
            },
...
            "UpdateAlertRequest": {
                "description": "Validación para la actualización de alertas",
                "required": [
                    "tipo",
                    "fecha_hora",
                    "descripcion",
                    "completado",
                    "categoria_id",
                    "paciente_id",
                    "operador_id"
                ],
                "properties": {
                    "id": {
                        "description": "ID de la alerta existente",
                        "type": "integer",
                        "example": 1,
                        "nullable": true
                    },
                    "tipo": {
                        "description": "Tipo de alerta",
                        "type": "string",
                        "enum": [
                            "puntual",
                            "periodico"
                        ],
                        "example": "puntual"
                    },
                    "fecha_hora": {
                        "description": "Fecha y hora de la alerta",
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-02-24T15:30:00Z"
                    },
                    "descripcion": {
                        "description": "Descripción de la alerta",
                        "type": "string",
                        "example": "Revisión médica programada"
                    },
                    "completado": {
                        "description": "Indica si la alerta ha sido completada",
                        "type": "boolean",
                        "example": true
                    },
                    "fecha_completado": {
                        "description": "Fecha y hora en que se completó la alerta",
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-02-25T10:00:00Z",
                        "nullable": true
                    },
                    "categoria_id": {
                        "description": "ID de la categoría de la alerta",
                        "type": "integer",
                        "example": 3
                    },
                    "paciente_id": {
                        "description": "ID del paciente asociado a la alerta",
                        "type": "integer",
                        "example": 5
                    },
                    "operador_id": {
                        "description": "ID del operador encargado de la alerta",
                        "type": "integer",
                        "example": 2
                    }
                },
                "type": "object"
            },
...
            "AvisoResource": {
                "description": "Esquema del recurso Aviso",
                "properties": {
                    "id": {
                        "description": "Identificador del aviso",
                        "type": "integer",
                        "example": 1
                    },
                    "tipo": {
                        "description": "Tipo de aviso",
                        "type": "string",
                        "enum": [
                            "puntual",
                            "periodico"
                        ],
                        "example": "puntual"
                    },
                    "fecha_hora": {
                        "description": "Fecha y hora programada para el aviso",
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-02-24 15:30:00"
                    },
                    "descripcion": {
                        "description": "Descripción del aviso",
                        "type": "string",
                        "example": "Revisión médica programada para el paciente"
                    },
                    "completado": {
                        "description": "Indica si el aviso ha sido completado",
                        "type": "boolean",
                        "example": true
                    },
                    "fecha_completado": {
                        "description": "Fecha y hora en que se completó el aviso",
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-02-25 10:00:00",
                        "nullable": true
                    },
                    "categoria": {
                        "$ref": "#/components/schemas/CategoriaResource"
                    },
                    "paciente": {
                        "$ref": "#/components/schemas/PatientResource"
                    },
                    "operador": {
                        "$ref": "#/components/schemas/UserResource"
                    },
                    "zona": {
                        "$ref": "#/components/schemas/ZonaResource"
                    },
                    "created_at": {
                        "description": "Fecha de creación del aviso",
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-02-23 12:00:00"
                    },
                    "updated_at": {
                        "description": "Fecha de última actualización del aviso",
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-02-24 12:00:00"
                    }
                },
                "type": "object"
            },
...
            "CategoriaResource": {
                "description": "Esquema del recurso Categoría",
                "properties": {
                    "id": {
                        "description": "Identificador de la categoría",
                        "type": "integer",
                        "example": 1
                    },
                    "nombre": {
                        "description": "Nombre de la categoría",
                        "type": "string",
                        "example": "Urgencias Médicas"
                    },
                    "descripcion": {
                        "description": "Descripción de la categoría",
                        "type": "string",
                        "example": "Categoría para gestionar urgencias médicas prioritarias",
                        "nullable": true
                    },
                    "subcategorias": {
                        "description": "Lista de subcategorías asociadas a la categoría",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/SubcategoriaResource"
                        }
                    },
                    "created_at": {
                        "description": "Fecha de creación de la categoría",
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-02-23 12:00:00"
                    },
                    "updated_at": {
                        "description": "Fecha de última actualización de la categoría",
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-02-24 12:00:00"
                    }
                },
                "type": "object"
            },
...
            "PatientResource": {
                "description": "Esquema del recurso Paciente",
                "properties": {
                    "id": {
                        "description": "Identificador único del paciente",
                        "type": "integer",
                        "example": 1
                    },
                    "nombre": {
                        "description": "Nombre completo del paciente",
                        "type": "string",
                        "example": "María López"
                    },
                    "fecha_nacimiento": {
                        "description": "Fecha de nacimiento del paciente",
                        "type": "string",
                        "format": "date",
                        "example": "1985-04-12"
                    },
                    "direccion": {
                        "description": "Dirección del paciente",
                        "type": "string",
                        "example": "Calle Falsa 123, Madrid"
                    },
                    "ciudad": {
                        "description": "Ciudad de residencia del paciente",
                        "type": "string",
                        "example": "Madrid",
                        "nullable": true
                    },
                    "codigo_postal": {
                        "description": "Código postal del paciente",
                        "type": "string",
                        "maxLength": 10,
                        "example": "28080",
                        "nullable": true
                    },
                    "dni": {
                        "description": "DNI único del paciente",
                        "type": "string",
                        "maxLength": 20,
                        "example": "12345678A"
                    },
                    "tarjeta_sanitaria": {
                        "description": "Número de tarjeta sanitaria único",
                        "type": "string",
                        "maxLength": 50,
                        "example": "TS1234567890"
                    },
                    "telefono": {
                        "description": "Número de teléfono del paciente",
                        "type": "string",
                        "maxLength": 20,
                        "example": "+34 612 345 678"
                    },
                    "email": {
                        "description": "Correo electrónico del paciente",
                        "type": "string",
                        "format": "email",
                        "maxLength": 100,
                        "example": "maria.lopez@example.com",
                        "nullable": true
                    },
                    "zona_id": {
                        "description": "ID de la zona asignada al paciente",
                        "type": "integer",
                        "example": 3
                    },
                    "situacion_personal": {
                        "description": "Situación personal del paciente",
                        "type": "string",
                        "example": "Vive solo",
                        "nullable": true
                    },
                    "estado_salud": {
                        "description": "Estado de salud actual del paciente",
                        "type": "string",
                        "example": "Hipertensión",
                        "nullable": true
                    },
                    "condicion_vivienda": {
                        "description": "Condición de la vivienda del paciente",
                        "type": "string",
                        "example": "Vivienda propia en buen estado",
                        "nullable": true
                    },
                    "nivel_autonomia": {
                        "description": "Nivel de autonomía del paciente",
                        "type": "string",
                        "example": "Autónomo",
                        "nullable": true
                    },
                    "situacion_economica": {
                        "description": "Situación económica del paciente",
                        "type": "string",
                        "example": "Media-baja",
                        "nullable": true
                    },
                    "created_at": {
                        "description": "Fecha de creación del paciente",
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-02-23 12:00:00"
                    },
                    "updated_at": {
                        "description": "Fecha de última actualización del paciente",
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-02-24 12:00:00"
                    }
                },
                "type": "object"
            },
            "SubcategoriaResource": {
                "description": "Esquema del recurso Subcategoría",
                "properties": {
                    "id": {
                        "description": "Identificador único de la subcategoría",
                        "type": "integer",
                        "example": 1
                    },
                    "nombre": {
                        "description": "Nombre de la subcategoría",
                        "type": "string",
                        "example": "Emergencia leve"
                    },
                    "categoria_id": {
                        "$ref": "#/components/schemas/CategoriaResource"
                    },
                    "created_at": {
                        "description": "Fecha de creación de la subcategoría",
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-02-23 12:00:00"
                    },
                    "updated_at": {
                        "description": "Fecha de última actualización de la subcategoría",
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-02-24 12:00:00"
                    }
                },
                "type": "object"
            },
            "UserResource": {
                "description": "Esquema del recurso Usuario",
                "properties": {
                    "id": {
                        "description": "Identificador único del usuario",
                        "type": "integer",
                        "example": 1
                    },
                    "nombre": {
                        "description": "Nombre completo del usuario",
                        "type": "string",
                        "example": "Juan Pérez"
                    },
                    "email": {
                        "description": "Correo electrónico del usuario",
                        "type": "string",
                        "format": "email",
                        "example": "juan.perez@example.com"
                    },
                    "role": {
                        "description": "Rol asignado al usuario",
                        "type": "string",
                        "example": "admin"
                    },
                    "is_active": {
                        "description": "Indica si el usuario está activo",
                        "type": "boolean",
                        "example": true
                    },
                    "zona_id": {
                        "description": "ID de la zona asignada al usuario",
                        "type": "integer",
                        "example": 3
                    },
                    "created_at": {
                        "description": "Fecha de creación del usuario",
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-02-23 12:00:00"
                    },
                    "updated_at": {
                        "description": "Fecha de última actualización del usuario",
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-02-24 12:00:00"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}